refactor(rust): extract babylon-tick run_once seam for the client (B0)
Move the Slice 1 flow (scenario load -> load_rule with LoadContext ->
pre-hash -> run_tick -> post-hash) out of main.rs into a new
babylon_tick::run_once(scenario_src, rule_src) -> Result<TickReport, String>
library function, with TickReport { before, after, fired } and the hex()
formatter both exported. main.rs is now argument parsing, reading the two
files, calling run_once, and printing.

This is the seam babylon-client's engine link (Task 12) calls in-process —
"the client links the engine" means literally sharing this code path.

Verified byte-identical: before/after hashes on two-classes.bscn +
fundamental-theorem.bsl are unchanged from the pre-refactor CLI output
(5a44ab0c...85a205 / 783f651d...427679), matching the plan's pinned golden
exactly — no delta to record.
